+++ braindecode/training/__init__.py
@@ -5,4 +5,4 @@ Functionality for skorch-based training.
 
 from .losses import CroppedLoss, mixup_criterion, TimeSeriesLoss
 from .scoring import (CroppedTrialEpochScoring, PostEpochTrainScoring,
-                      trial_preds_from_window_preds, predict_trials)
+                      CroppedTimeSeriesEpochScoring, trial_preds_from_window_preds, predict_trials)